+++ v1/back/java/src/main/java/fr/gouv/monprojetsup/data/ServerData.java
@@ -8,7 +8,6 @@ import fr.gouv.monprojetsup.data.model.formations.Formation;
 import fr.gouv.monprojetsup.data.model.specialites.Specialites;
 import fr.gouv.monprojetsup.data.model.specialites.SpecialitesLoader;
 import fr.gouv.monprojetsup.data.model.stats.PsupStatistiques;
-import fr.gouv.monprojetsup.data.model.stats.Statistique;
 import fr.gouv.monprojetsup.data.model.stats.StatsContainers;
 import fr.gouv.monprojetsup.data.update.BackEndData;
 import fr.gouv.monprojetsup.data.update.onisep.DomainePro;
@@ -17,7 +16,6 @@ import fr.gouv.monprojetsup.data.update.psup.PsupData;
 import fr.gouv.monprojetsup.suggestions.algos.Explanation;
 import fr.gouv.monprojetsup.tools.Serialisation;
 import org.apache.commons.lang3.tuple.Pair;
-import org.apache.commons.lang3.tuple.Triple;
 import org.apache.commons.text.similarity.LevenshteinDistance;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
