--- v1/back/java/src/main/java/fr/gouv/monprojetsup/tools/expeEns/ImportGroupsFromCsv.java
@@ -1,39 +1,0 @@
-package fr.gouv.monprojetsup.tools.expeEns;
-
-import fr.gouv.monprojetsup.web.db.DBExceptions;
-import fr.gouv.monprojetsup.web.db.dbimpl.DBMongo;
-import fr.gouv.monprojetsup.web.server.WebServerConfig;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.WebApplicationType;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-import java.io.IOException;
-
-@SpringBootApplication
-@EnableMongoRepositories
-@ComponentScan(basePackages = {"fr.gouv.monprojetsup"})
-public class ImportGroupsFromCsv {
-
-    public static void main(String[] args) throws IOException, DBExceptions.ModelException {
-
-        WebServerConfig config = WebServerConfig.load();
-
-        ConfigurableApplicationContext context =
-                new SpringApplicationBuilder(ImportGroupsFromCsv.class)
-                        .web(WebApplicationType.NONE)
-                        .run(args);
-
-        DBMongo db = context.getBean(DBMongo.class);
-
-        db.load(config);
-
-        db.injectFromCsvFile(config.listeClassesCsvPath);
-
-        SpringApplication.exit(context);
-    }
-}
-
